refactor: remove map from parameter and return value of GetSecrets (#62)
* refactor: remove map from parameter and return value of GetSecrets

This PR refators the functin addon.GetSecrets that previously required
us to build a map of endpoint <-> secretRef and then return a map of
endpoint <-> secret. This map made sense when the addon had to generate
its secrets for CLF/OTELCol. Since this design was changed when we
introduced clustersets we stoped using the endpoint key in the code. The
addon.GetSecrets function now only required us to provide a list of
Secrets.

* Apply suggestions from code review

Co-authored-by: Periklis Tsirakidis <periklis@nefeli.eu>

* review comment don't pre-allocate

---------

Co-authored-by: Periklis Tsirakidis <periklis@nefeli.eu>